refactor(ui): Move repository table to product main page
To improve usability by saving extra clicks, show all repositories 
belonging to a product in the product main page. In top-right, the total
number of repositories is shown, along with a button to create a
repository.
